Move commpn test classes to TestCore
